+++ src/Har/HarHandler.php
@@ -19,34 +19,29 @@ use Berlioz\Http\Client\Exception\HttpClientException;
 use Berlioz\Http\Client\History\HistoryEntry;
 use Berlioz\Http\Client\History\Timings;
 use Berlioz\Http\Client\Session;
-use ElGigi\HarParser\Entities\Entry;
-use ElGigi\HarParser\Entities\Header;
-use ElGigi\HarParser\Entities\Log;
-use ElGigi\HarParser\Entities\Request;
-use ElGigi\HarParser\Entities\Response;
+use Berlioz\Http\Message\Request;
+use Berlioz\Http\Message\Response;
+use ElGigi\HarParser\Entities as Har;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
 
-/**
- * Class HarParser.
- */
-class HarParser
+class HarHandler
 {
     /**
      * @throws HttpClientException
      */
-    public function handle(Log $har): Session
+    public function handle(Har\Log $har): Session
     {
         $session = new Session();
 
-        /** @var Entry|null $entry */
+        /** @var Har\Entry|null $entry */
         if ($entry = $har->getEntries()->current()) {
             foreach ($entry->getRequest()->getCookies() as $cookie) {
                 $session->getCookies()->addCookie(Cookie::createFromHar($cookie));
             }
         }
 
-        /** @var Entry $entry */
+        /** @var Har\Entry $entry */
         foreach ($har->getEntries() as $entry) {
             $this->addEntryToSession($session, $entry);
         }
@@ -58,11 +53,11 @@ class HarParser
      * Add entry to session.
      *
      * @param Session $session
-     * @param Entry $entry
+     * @param Har\Entry $entry
      *
      * @throws HttpClientException
      */
-    public function addEntryToSession(Session $session, Entry $entry): void
+    public function addEntryToSession(Session $session, Har\Entry $entry): void
     {
         $request = $this->getHttpRequest($entry->getRequest());
         $response = $this->getHttpResponse($entry->getResponse());
@@ -75,17 +70,32 @@ class HarParser
     /**
      * Get HTTP request.
      *
-     * @param Request $request
+     * @param Har\Request $request
      *
      * @return RequestInterface
      */
-    public function getHttpRequest(Request $request): RequestInterface
+    public function getHttpRequest(Har\Request $request): RequestInterface
     {
-        $httpRequest = new \Berlioz\Http\Message\Request(
+        $httpRequest = new Request(
             method: $request->getMethod(),
             uri: $request->getUrl(),
             body: $request->getPostData()?->getText(),
-            headers: $this->getHeaders($request->getHeaders()),
+            headers: $this->getHeaders($request),
+        );
+
+        // Cookies from HAR entry
+        // Do not use the header defined in entry, because can be anonymized!
+        $httpRequest = $httpRequest->withHeader(
+            'Cookie',
+            [
+                implode(
+                    '; ',
+                    array_map(
+                        fn(Har\Cookie $cookie) => Cookie::createFromHar($cookie)->getRequestHeader(),
+                        $request->getCookies()
+                    )
+                )
+            ]
         );
 
         return $httpRequest->withProtocolVersion($request->getHttpVersion());
@@ -94,22 +104,32 @@ class HarParser
     /**
      * Get HTTP response.
      *
-     * @param Response $response
+     * @param Har\Response $response
      *
      * @return ResponseInterface
      */
-    public function getHttpResponse(Response $response): ResponseInterface
+    public function getHttpResponse(Har\Response $response): ResponseInterface
     {
         $body = $response->getContent()->getText();
         if ('base64' === $response->getContent()->getEncoding()) {
             $body = base64_decode($body);
         }
 
-        $httpResponse = new \Berlioz\Http\Message\Response(
+        $httpResponse = new Response(
             body: $body,
             statusCode: $response->getStatus(),
-            headers: $this->getHeaders($response->getHeaders()),
+            headers: $this->getHeaders($response),
             reasonPhrase: $response->getStatusText(),
+        );
+
+        // Cookies from HAR entry
+        // Do not use the header defined in entry, because can be anonymized!
+        $httpResponse = $httpResponse->withHeader(
+            'Set-Cookie',
+            array_map(
+                fn(Har\Cookie $cookie) => Cookie::createFromHar($cookie)->getRequestHeader(),
+                $response->getCookies()
+            )
         );
 
         return $httpResponse->withProtocolVersion($response->getHttpVersion());
@@ -118,11 +138,11 @@ class HarParser
     /**
      * Get timings.
      *
-     * @param Entry $entry
+     * @param Har\Entry $entry
      *
      * @return Timings
      */
-    public function getTimings(Entry $entry): Timings
+    public function getTimings(Har\Entry $entry): Timings
     {
         return new Timings(
             dateTime: $entry->getStartedDateTime(),
@@ -137,12 +157,12 @@ class HarParser
         );
     }
 
-    protected function getHeaders(array $headers): array
+    protected function getHeaders(Har\Message $message): array
     {
         $final = [];
 
-        /** @var Header $header */
-        foreach ($headers as $header) {
+        /** @var Har\Header $header */
+        foreach ($message->getHeaders() as $header) {
             $final[$header->getName()][] = $header->getValue();
         }